# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Qoyod Vendor & Purchase Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  email: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  admin: 'admin',
  accounts_manager: 'accounts_manager',
  accounts_employee: 'accounts_employee',
  employee: 'employee',
} as const;

export interface AuthUser {
  email: string;
  role: AuthUserRole;
  /** @nullable */
  name?: string | null;
}

export interface VendorCategory {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface VendorCategoryInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface VendorDocument {
  id: number;
  vendorId: number;
  documentType: string;
  /** @nullable */
  documentNumber?: string | null;
  /** @nullable */
  expiryDate?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface Vendor {
  id: number;
  companyName: string;
  /** @nullable */
  contactPerson?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  bankAccountName?: string | null;
  /** @nullable */
  bankAccountNumber?: string | null;
  /** @nullable */
  iban?: string | null;
  /** @nullable */
  swiftCode?: string | null;
  /** @nullable */
  bankBranch?: string | null;
  /** @nullable */
  notes?: string | null;
  categoryIds?: number[];
  categories?: VendorCategory[];
  documents?: VendorDocument[];
  /** @nullable */
  totalSpent?: number | null;
  /** @nullable */
  transactionCount?: number | null;
  createdAt: string;
}

export interface VendorInput {
  /** @minLength 1 */
  companyName: string;
  contactPerson?: string;
  contactEmail?: string;
  contactPhone?: string;
  bankName?: string;
  bankAccountName?: string;
  bankAccountNumber?: string;
  iban?: string;
  swiftCode?: string;
  bankBranch?: string;
  notes?: string;
  categoryIds?: number[];
}

export interface VendorDashboard {
  totalVendors: number;
  totalSpent: number;
  activeVendors: number;
  expiringDocumentsCount?: number;
}

export interface VendorImport {
  csvData: string;
}

export interface ImportResult {
  imported: number;
  errors: string[];
}

export interface VendorDocumentInput {
  /** @minLength 1 */
  documentType: string;
  documentNumber?: string;
  expiryDate?: string;
  fileUrl?: string;
  notes?: string;
}

export interface ExpiringDocument {
  vendorId: number;
  vendorName: string;
  documentType: string;
  expiryDate: string;
  daysUntilExpiry: number;
}

export type PurchaseRequestStatus = typeof PurchaseRequestStatus[keyof typeof PurchaseRequestStatus];


export const PurchaseRequestStatus = {
  pending_manager: 'pending_manager',
  pending_clarification_employee_manager: 'pending_clarification_employee_manager',
  pending_clarification_employee_accounts: 'pending_clarification_employee_accounts',
  approved_by_manager: 'approved_by_manager',
  rejected_by_manager: 'rejected_by_manager',
  pending_accounts: 'pending_accounts',
  approved_by_accounts: 'approved_by_accounts',
  rejected_by_accounts: 'rejected_by_accounts',
  executed: 'executed',
} as const;

export interface PurchaseRequest {
  id: number;
  requestNumber: string;
  requesterEmail: string;
  department: string;
  itemDescription: string;
  quantity: number;
  vendorId: number;
  vendor?: Vendor;
  reason: string;
  managerEmail: string;
  status: PurchaseRequestStatus;
  /** @nullable */
  estimatedAmount?: number | null;
  /** @nullable */
  finalAmount?: number | null;
  /** @nullable */
  managerNote?: string | null;
  /** @nullable */
  accountsNote?: string | null;
  /** @nullable */
  clarificationQuestion?: string | null;
  /** @nullable */
  clarificationAnswer?: string | null;
  /** @nullable */
  executedAt?: string | null;
  /** @nullable */
  executedBy?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface PurchaseRequestInput {
  requesterEmail: string;
  /** @minLength 1 */
  department: string;
  /** @minLength 1 */
  itemDescription: string;
  /** @minimum 1 */
  quantity: number;
  vendorId: number;
  /** @minLength 1 */
  reason: string;
  managerEmail: string;
  estimatedAmount?: number;
}

export interface ActionInput {
  note?: string;
}

export interface ClarificationResponse {
  /** @minLength 1 */
  answer: string;
  updatedQuantity?: number;
  updatedReason?: string;
  updatedManagerEmail?: string;
}

export interface ExecuteInput {
  executedByEmail: string;
  finalAmount: number;
  notes?: string;
}

export interface ReorderInput {
  requesterEmail: string;
  /** @minimum 1 */
  quantity: number;
  reason: string;
  managerEmail: string;
  estimatedAmount?: number;
}

export interface RequestActivity {
  id: number;
  requestId: number;
  actorEmail: string;
  action: string;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalRequests: number;
  pendingRequests: number;
  approvedRequests: number;
  rejectedRequests: number;
  executedRequests: number;
  totalVendors: number;
  totalSpent: number;
  expiringDocuments?: number;
}

export interface VendorSpending {
  vendorId: number;
  vendorName: string;
  totalSpent: number;
  transactionCount: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export type ReportDataFilters = {
  /** @nullable */
  vendorId?: number | null;
  /** @nullable */
  month?: string | null;
  type?: string;
};

export interface ReportData {
  requests: PurchaseRequest[];
  totalAmount: number;
  generatedAt: string;
  filters?: ReportDataFilters;
}

export type ListVendorsParams = {
categoryId?: number;
search?: string;
};

export type ListPurchaseRequestsParams = {
status?: string;
vendorId?: number;
requesterEmail?: string;
managerEmail?: string;
};

export type GetReportParams = {
vendorId?: number;
/**
 * YYYY-MM
 */
month?: string;
type?: GetReportType;
};

export type GetReportType = typeof GetReportType[keyof typeof GetReportType];


export const GetReportType = {
  summary: 'summary',
  detailed: 'detailed',
} as const;