# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Qoyod Vendor & Purchase Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /purchase-requests:
    get:
      operationId: listPurchaseRequests
      tags: [purchase-requests]
      summary: List purchase requests (filtered by role)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: requesterEmail
          in: query
          schema:
            type: string
        - name: managerEmail
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of purchase requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseRequest"
    post:
      operationId: createPurchaseRequest
      tags: [purchase-requests]
      summary: Submit a new purchase request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
  /purchase-requests/{id}:
    get:
      operationId: getPurchaseRequest
      tags: [purchase-requests]
      summary: Get a purchase request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /purchase-requests/{id}/approve:
    post:
      operationId: approvePurchaseRequest
      tags: [purchase-requests]
      summary: Approve a purchase request (manager or accounts manager)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionInput"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
  /purchase-requests/{id}/reject:
    post:
      operationId: rejectPurchaseRequest
      tags: [purchase-requests]
      summary: Reject a purchase request (manager or accounts manager)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
  /purchase-requests/{id}/clarify:
    post:
      operationId: clarifiyPurchaseRequest
      tags: [purchase-requests]
      summary: Request clarification from employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionInput"
      responses:
        "200":
          description: Clarification requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
  /purchase-requests/{id}/respond:
    post:
      operationId: respondToClarification
      tags: [purchase-requests]
      summary: Employee responds to clarification and resubmits
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClarificationResponse"
      responses:
        "200":
          description: Resubmitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
  /purchase-requests/{id}/execute:
    post:
      operationId: executePurchaseRequest
      tags: [purchase-requests]
      summary: Accounts employee marks request as executed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteInput"
      responses:
        "200":
          description: Executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
  /purchase-requests/{id}/reorder:
    post:
      operationId: reorderPurchaseRequest
      tags: [purchase-requests]
      summary: Reorder from same vendor (creates new request based on existing)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "201":
          description: New request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
  /purchase-requests/{id}/activities:
    get:
      operationId: listPurchaseRequestActivities
      tags: [purchase-requests]
      summary: Get activity/comment history for a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequestActivity"

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Auth ───────────────────────────────────────────────────────────────
    LoginInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    AuthUser:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
        role:
          type: string
          enum: [admin, accounts_manager, accounts_employee, employee]
        name:
          type: ["string", "null"]

    # ── Vendor Category ────────────────────────────────────────────────────
    VendorCategory:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Vendor:
      type: object
      required: [id, companyName, createdAt]
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactPerson:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        bankAccountName:
          type: ["string", "null"]
        bankAccountNumber:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
        swiftCode:
          type: ["string", "null"]
        bankBranch:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        categoryIds:
          type: array
          items:
            type: integer
        categories:
          type: array
          items:
            $ref: "#/components/schemas/VendorCategory"
        documents:
          type: array
          items:
            $ref: "#/components/schemas/VendorDocument"
        totalSpent:
          type: ["number", "null"]
        transactionCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    VendorDocument:
      type: object
      required: [id, vendorId, documentType, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        documentType:
          type: string
        documentNumber:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
          format: date
        fileUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    PurchaseRequest:
      type: object
      required: [id, requestNumber, requesterEmail, department, itemDescription, quantity, vendorId, reason, managerEmail, status, createdAt]
      properties:
        id:
          type: integer
        requestNumber:
          type: string
        requesterEmail:
          type: string
        department:
          type: string
        itemDescription:
          type: string
        quantity:
          type: integer
        vendorId:
          type: integer
        vendor:
          $ref: "#/components/schemas/Vendor"
        reason:
          type: string
        managerEmail:
          type: string
        status:
          type: string
          enum:
            - pending_manager
            - pending_clarification_employee_manager
            - pending_clarification_employee_accounts
            - approved_by_manager
            - rejected_by_manager
            - pending_accounts
            - approved_by_accounts
            - rejected_by_accounts
            - executed
        estimatedAmount:
          type: ["number", "null"]
        finalAmount:
          type: ["number", "null"]
        managerNote:
          type: ["string", "null"]
        accountsNote:
          type: ["string", "null"]
        clarificationQuestion:
          type: ["string", "null"]
        clarificationAnswer:
          type: ["string", "null"]
        executedAt:
          type: ["string", "null"]
          format: date-time
        executedBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PurchaseRequestInput:
      type: object
      required: [requesterEmail, department, itemDescription, quantity, vendorId, reason, managerEmail]
      properties:
        requesterEmail:
          type: string
          format: email
        department:
          type: string
          minLength: 1
        itemDescription:
          type: string
          minLength: 1
        quantity:
          type: integer
          minimum: 1
        vendorId:
          type: integer
        reason:
          type: string
          minLength: 1
        managerEmail:
          type: string
          format: email
        estimatedAmount:
          type: number
    ActionInput:
      type: object
      properties:
        note:
          type: string
    ClarificationResponse:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
          minLength: 1
        updatedQuantity:
          type: integer
        updatedReason:
          type: string
        updatedManagerEmail:
          type: string
    ExecuteInput:
      type: object
      required: [executedByEmail, finalAmount]
      properties:
        executedByEmail:
          type: string
          format: email
        finalAmount:
          type: number
        notes:
          type: string
    ReorderInput:
      type: object
      required: [requesterEmail, quantity, reason, managerEmail]
      properties:
        requesterEmail:
          type: string
          format: email
        quantity:
          type: integer
          minimum: 1
        reason:
          type: string
        managerEmail:
          type: string
          format: email
        estimatedAmount:
          type: number
    RequestActivity:
      type: object
      required: [id, requestId, actorEmail, action, createdAt]
      properties:
        id:
          type: integer
        requestId:
          type: integer
        actorEmail:
          type: string
        action:
          type: string
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ── Dashboard ──────────────────────────────────────────────────────────
...